Move db config
